Cost Center Updates

+++ GVWebApi/Models/CostAllocation/CostAllocationSettingsViewModel.cs
@@ -9,7 +9,7 @@ namespace GVWebapi.Models
     {
         public int SettingsMeterGroupID { get; set; }
         public int CustomerID { get; set; }
-        public int MeterGroupID { get; set; }
+        public long MeterGroupID { get; set; }
         public string MeterGroupDesc { get; set; }
         public decimal ExcessCPP { get; set; }
     }
